+++ src/types/options/TInputOptions.ts
@@ -1,5 +1,8 @@
 import { TDictionary } from '../TDictionary';
+import { TInputOptionsPreset } from './TInputOptionsPreset';
 
 import { IOptions } from '../../interfaces/options/IOptions';
 
-export type TInputOptions = Partial<Pick<IOptions, keyof IOptions>> & TDictionary;
+export type TInputOptions = Partial<Omit<IOptions, 'optionsPreset'>> & {
+    optionsPreset?: TInputOptionsPreset;
+} & TDictionary;